Move audio media block to its own file
