+++ di_website/templates/publications/audio_visual_media.html
@@ -40,7 +40,6 @@
 
     <section class="section">
       <div class="row row--narrow">
-        <h1 class="audio-player-title">Listen to audio</h1>
         <div class="l-sidebar">
           {{page.media_content}}
         </div>
